move style file to appease next build tool
